[PATCH] [PATCH] fix a FIXME in usb.h
In ush.h, there's a FIXME for the URB transfer flags. This patch is
basically a global search and replace to change those all from USB_ to
URB_.

It touches a few things that aren't directly USB-related, and so should
probably be passed by those authors, but I figured i should put it here to
get feedback (ie: "No, moron, you did it all wrong!" or "Oops, that FIXME
wasn't supposed to be there") before bothering them.