[PATCH] 2.5.45 CompactPCI driver patch 4/4
This is a patch 4 of 4 of my CompactPCI hotplug core and
drivers, consisting of the generic port I/O cPCI hotplug driver.

Let me know if the kernel parameter parsing code that's #ifndef MODULE is
objectionable.  I spent quite a while today testing it, it seems reasonably
robust.  Without it, this driver would only be useable as a module, which
I've not figured out how to do with the new kernel configuration stuff.