[PATCH] `event' removal: other filesystems
Patch from Manfred Spraul

Several filesystems compare f_version and i_version to validate
directory positions in readdir(): The directory position is revalidated
if i_version is not equal f_version.  Operations that could invalidate
the cached position set i_version or f_version to '++event', event is a
global variable.  Global uniqueness is not needed, 'i_version++' and
'f_version=0' is sufficient to guarantee that the next readdir() will
revalidate the directory position, and that avoids the need for an ugly
global variable.

The attached patch converts all filesystems except ext2, which was
converted with a seperate patch.