sysfs: fix file deletion again.
After looking into again, I realize that I was again getting file and 
directory deletion a little wrong. This patch should go back to mimmicking
the VFS again. 

- The extra dget() in sysfs_mknod() and sysfs_symlink() has been readded. 
This is identical to the way ramfs does creation. 

- We do d_delete() after simple_unlink() and simple_rmdir(), instead of 
d_invalidate() before it. This is how vfs_rmdir() and vfs_unlink() do it, 
and the way we used to.